+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add [`QuinticSplineMethod`] for accurate filament discretisation.
+
 ## [0.5.1] - 2023-10-24
 
 ### Added